restructured files and added data directly to the front end
