+++ client/src/App.jsx
@@ -1,31 +1,21 @@
-import { useState } from 'react';
 import { Route, BrowserRouter, Routes } from 'react-router-dom';
 import GlobalStyles from './GlobalStyles';
-import Navbar from './Navbar';
-import Wallets from './Wallets';
-import Footer from './Footer';
-import { UserProvider } from './UserContext';
-import Signin from './Signin';
-import Profile from './Profile';
-import Signup from './Signup';
+import Navbar from './components/Navbar';
+import Wallets from './components/Wallets';
+import Footer from './components/Footer';
 
 
 function App() {
 
   return (
-    <UserProvider>
       <BrowserRouter>
       <GlobalStyles />
         <Navbar />
         <Routes>
           <Route path='/' element={<Wallets />} />
-          <Route path="/signin" element={<Signin />} />
-          <Route path="/profile" element={<Profile />} />
-          <Route path="/signup" element={<Signup />} />
         </Routes>
         <Footer />
       </BrowserRouter>
-    </UserProvider>
   );
 }
