+++ client/src/components/Navbar.jsx
@@ -1,16 +1,10 @@
 import React, { useState, useEffect, useContext } from 'react';
 import styled from 'styled-components';
-import UserContext, { useUser } from './UserContext';
 import { Link } from 'react-router-dom';
-import blockIcon from './assets/block-01.svg';
+import blockIcon from '../assets/block-01.svg';
 
 const Navbar = () => {
     const [blockHeight, setBlockHeight] = useState(null);
-    const { user, setUser } = useContext(UserContext);
-
-    const handleSignout = () => {
-        setUser(null);
-    };
 
     useEffect(() => {
         // Fetch the current Bitcoin block height
@@ -26,16 +20,6 @@ const Navbar = () => {
             <BlockContainer>
                 <Icon src={blockIcon} alt="Block Icon"/> Block: {blockHeight}
             </BlockContainer>
-            <LinkContainer>
-                {!user ? (
-                <Link to="/signin">Sign In</Link>
-            ) : (
-                <>
-                    <Link to="/profile">Profile</Link>
-                    <button onClick={handleSignout}>Sign Out</button>
-                </>
-            )}
-            </LinkContainer>
         </NavContainer>
     );
 };
@@ -66,24 +50,5 @@ const Icon = styled.img`
     padding: 10px;
     width: 3vw;
 `;
-
-const ProfileIcon = styled.button`
-    background: none;
-    border: none;
-    color: #fff;
-    cursor: pointer;
-`;
-
-const LinkContainer = styled(Link)`
-    text-decoration: none;      
-    color: #fff;                
-    font-size: 15px;            
-    font-weight: bold;         
-    letter-spacing: 1px;        
-    text-transform: uppercase;
-    &:visited {  
-        color: #fff;
-    }
-`
 
 export default Navbar;
