--- client/src/components/UserContext.jsx
@@ -1,19 +1,0 @@
-import React, { createContext, useState, useContext } from 'react';
-
-const UserContext = createContext();
-
-export const UserProvider = ({ children }) => {
-    const [user, setUser] = useState(null);
-
-    return (
-        <UserContext.Provider value={{ user, setUser }}>
-            {children}
-        </UserContext.Provider>
-    );
-};
-
-export const useUser = () => {
-    return useContext(UserContext);
-};
-
-export default UserContext;
